+++ lib/src/controls/form/auto_suggest_box.dart
@@ -1,65 +1,67 @@
 import 'package:fluent_ui/fluent_ui.dart';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
+
+enum TextChangedReason {
+  /// Whether the text in an [AutoSuggestBox] was changed by user input
+  userInput,
+
+  /// Whether the text in an [AutoSuggestBox] was changed because the user
+  /// chose the suggestion
+  suggestionChosen,
+}
 
 // TODO: Navigate through items using keyboard (https://github.com/bdlukaa/fluent_ui/issues/19)
-
-typedef AutoSuggestBoxItemBuilder<T> = Widget Function(BuildContext, T);
-typedef AutoSuggestBoxItemSorter<T> = List<T> Function(String, List<T>);
-typedef AutoSuggestBoxTextBoxBuilder<T> = Widget Function(
-  BuildContext context,
-  TextEditingController controller,
-  FocusNode focusNode,
-  GlobalKey key,
-);
 
 /// An AutoSuggestBox provides a list of suggestions for a user to select
 /// from as they type.
 ///
-/// ![AutoSuggestBox Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/controls_autosuggest_expanded01.png)
+/// ![AutoSuggestBox Preview](https://docs.microsoft.com/en-us/windows/apps/design/controls/images/controls-autosuggest-expanded-01.png)
 ///
 /// See also:
-///   - [TextBox]
-///   - [ComboBox]
+///
+///   * <https://docs.microsoft.com/en-us/windows/apps/design/controls/auto-suggest-box>
+///   * [TextBox], which is used by this widget to enter user text input
+///   * [Overlay], which is used to show the popup
 class AutoSuggestBox<T> extends StatefulWidget {
   /// Creates a fluent-styled auto suggest box.
   const AutoSuggestBox({
     Key? key,
-    required this.controller,
     required this.items,
-    this.itemBuilder,
-    this.sorter = defaultItemSorter,
-    this.noResultsFound = defaultNoResultsFound,
-    this.textBoxBuilder = defaultTextBoxBuilder,
+    this.controller,
+    this.onChanged,
     this.onSelected,
+    this.trailingIcon,
+    this.clearButtonEnabled = true,
+    this.placeholder,
   }) : super(key: key);
+
+  /// The list of items to display to the user to pick
+  final List<T> items;
 
   /// The controller used to have control over what to show on
   /// the [TextBox].
-  final TextEditingController controller;
+  final TextEditingController? controller;
 
-  /// The list of items to display to the user to pick. If empty,
-  /// [noResultsFound] is used.
-  final List<T> items;
-
-  /// The item builder to build [items]. If null, uses a default
-  /// internal builder
-  final AutoSuggestBoxItemBuilder<T>? itemBuilder;
-
-  /// Sort the items to show. [defaultItemSorter] is used by default
-  final AutoSuggestBoxItemSorter sorter;
-
-  /// Build the text box. [defaultTextBoxBuilder] is used by default
-  final AutoSuggestBoxTextBoxBuilder<T> textBoxBuilder;
-
-  /// The widget to show when the text the user typed doesn't match with
-  /// [items]s. [defaultNoResultsFound] is used by default.
-  ///
-  /// ![No results found Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/controls_autosuggest_noresults.png)
-  final WidgetBuilder noResultsFound;
+  /// Called when the text is updated
+  final void Function(String text, TextChangedReason reason)? onChanged;
 
   /// Called when the user selected a value.
   final ValueChanged<T>? onSelected;
+
+  /// A widget displayed in the end of the [TextBox]
+  ///
+  /// Usually an [IconButton] or [Icon]
+  final Widget? trailingIcon;
+
+  /// Whether the close button is enabled
+  ///
+  /// Defauls to true
+  final bool clearButtonEnabled;
+
+  /// The placeholder
+  final String? placeholder;
 
   @override
   _AutoSuggestBoxState<T> createState() => _AutoSuggestBoxState<T>();
@@ -80,36 +82,6 @@ class AutoSuggestBox<T> extends StatefulWidget {
       return element.toString().toLowerCase().contains(text.toLowerCase());
     }).toList();
   }
-
-  /// Creates a 'No results found' tile.
-  ///
-  /// ![No results found Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/controls_autosuggest_noresults.png)
-  static Widget defaultNoResultsFound(context) {
-    return const ListTile(
-      title: DefaultTextStyle(
-        style: TextStyle(fontWeight: FontWeight.normal),
-        child: Text('No results found'),
-      ),
-    );
-  }
-
-  static Widget defaultTextBoxBuilder(
-    BuildContext context,
-    TextEditingController controller,
-    FocusNode focusNode,
-    GlobalKey key,
-  ) {
-    assert(debugCheckHasFluentLocalizations(context));
-    final FluentLocalizations localizations = FluentLocalizations.of(context);
-    return TextBox(
-      key: key,
-      controller: controller,
-      focusNode: focusNode,
-      placeholder: localizations.searchLabel,
-      clipBehavior:
-          focusNode.hasFocus ? Clip.none : Clip.antiAliasWithSaveLayer,
-    );
-  }
 }
 
 class _AutoSuggestBoxState<T> extends State<AutoSuggestBox<T>> {
@@ -118,47 +90,37 @@ class _AutoSuggestBoxState<T> extends State<AutoSuggestBox<T>> {
   final LayerLink _layerLink = LayerLink();
   final GlobalKey _textBoxKey = GlobalKey();
 
+  late TextEditingController controller;
+
   @override
   void initState() {
     super.initState();
+    controller = widget.controller ?? TextEditingController();
+    controller.addListener(() {
+      if (!mounted) return;
+      if (controller.text.length < 2) setState(() {});
+    });
     focusNode.addListener(_handleFocusChanged);
   }
 
   @override
   void dispose() {
     focusNode.removeListener(_handleFocusChanged);
+    if (widget.controller == null) {
+      controller.dispose();
+    }
     super.dispose();
   }
 
   void _handleFocusChanged() {
     final hasFocus = focusNode.hasFocus;
-    if (hasFocus) {
-      if (_entry == null && !(_entry?.mounted ?? false)) {
-        _insertOverlay();
-      }
-    } else {
+    if (!hasFocus) {
       _dismissOverlay();
     }
     setState(() {});
   }
 
-  AutoSuggestBoxItemBuilder<T> get itemBuilder =>
-      widget.itemBuilder ?? _defaultItemBuilder;
-
-  Widget _defaultItemBuilder(BuildContext context, T value) {
-    return TappableListTile(
-      onTap: () {
-        widget.controller.text = '$value';
-        widget.onSelected?.call(value);
-        focusNode.unfocus();
-      },
-      title:
-          Text('$value', style: FluentTheme.maybeOf(context)?.typography.body),
-    );
-  }
-
   void _insertOverlay() {
-    final acrylicDisabled = DisableAcrylic.of(context) != null;
     _entry = OverlayEntry(builder: (context) {
       final context = _textBoxKey.currentContext;
       if (context == null) return const SizedBox.shrink();
@@ -171,62 +133,25 @@ class _AutoSuggestBoxState<T> extends State<AutoSuggestBox<T>> {
           offset: Offset(0, box.size.height + 0.8),
           child: SizedBox(
             width: box.size.width,
-            child: Acrylic(
-              shape: RoundedRectangleBorder(
-                borderRadius: const BorderRadius.vertical(
-                  bottom: Radius.circular(4.0),
-                ),
-                side: BorderSide(
-                  color: FluentTheme.of(context).scaffoldBackgroundColor,
-                  width: 0.8,
-                ),
-              ),
-              child: ValueListenableBuilder<TextEditingValue>(
-                valueListenable: widget.controller,
-                builder: (context, value, _) {
-                  final items = widget.sorter(value.text, widget.items);
-                  late Widget result;
-                  if (items.isEmpty) {
-                    result = widget.noResultsFound(context);
-                  } else {
-                    result = ListView(
-                      shrinkWrap: true,
-                      children: List.generate(items.length, (index) {
-                        final item = items[index];
-                        return itemBuilder(context, item);
-                      }),
-                    );
-                  }
-                  return AnimatedSwitcher(
-                    duration: FluentTheme.of(context).fastAnimationDuration,
-                    switchInCurve: FluentTheme.of(context).animationCurve,
-                    transitionBuilder: (child, animation) {
-                      if (child is ListView) {
-                        return child;
-                      }
-                      return EntrancePageTransition(
-                        child: child,
-                        animation: animation,
-                        vertical: true,
-                      );
-                    },
-                    layoutBuilder: (child, children) =>
-                        child ?? const SizedBox(),
-                    child: result,
-                  );
-                },
-              ),
+            child: _AutoSuggestBoxOverlay(
+              controller: controller,
+              items: widget.items,
+              onSelected: (item) {
+                widget.onSelected?.call(item);
+                controller.text = item;
+                widget.onChanged?.call(item, TextChangedReason.userInput);
+              },
             ),
           ),
         ),
       );
 
-      if (acrylicDisabled) return DisableAcrylic(child: child);
       return child;
     });
 
     if (_textBoxKey.currentContext != null) {
       Overlay.of(context)?.insert(_entry!);
+      if (mounted) setState(() {});
     }
   }
 
@@ -238,13 +163,195 @@ class _AutoSuggestBoxState<T> extends State<AutoSuggestBox<T>> {
   @override
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
+    assert(debugCheckHasFluentLocalizations(context));
+
     return CompositedTransformTarget(
       link: _layerLink,
-      child: widget.textBoxBuilder(
-        context,
-        widget.controller,
-        focusNode,
-        _textBoxKey,
+      child: TextBox(
+        key: _textBoxKey,
+        controller: controller,
+        focusNode: focusNode,
+        placeholder: widget.placeholder,
+        clipBehavior: _entry != null ? Clip.none : Clip.antiAliasWithSaveLayer,
+        suffix: Row(children: [
+          if (widget.trailingIcon != null) widget.trailingIcon!,
+          if (widget.clearButtonEnabled && controller.text.isNotEmpty)
+            Padding(
+              padding: const EdgeInsets.only(left: 2.0),
+              child: IconButton(
+                icon: const Icon(FluentIcons.close),
+                onPressed: () {
+                  controller.clear();
+                  focusNode.unfocus();
+                },
+              ),
+            ),
+        ]),
+        suffixMode: OverlayVisibilityMode.always,
+        onChanged: (text) {
+          widget.onChanged?.call(text, TextChangedReason.userInput);
+          if (_entry == null && !(_entry?.mounted ?? false)) {
+            _insertOverlay();
+          }
+        },
+      ),
+    );
+  }
+}
+
+class _AutoSuggestBoxOverlay extends StatelessWidget {
+  const _AutoSuggestBoxOverlay({
+    Key? key,
+    required this.items,
+    required this.controller,
+    required this.onSelected,
+  }) : super(key: key);
+
+  final List items;
+  final TextEditingController controller;
+  final ValueChanged onSelected;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = FluentTheme.of(context);
+    final localizations = FluentLocalizations.of(context);
+    return FocusScope(
+      autofocus: true,
+      child: Container(
+        constraints: const BoxConstraints(
+          maxHeight: 385,
+        ),
+        decoration: ShapeDecoration(
+          shape: RoundedRectangleBorder(
+            borderRadius: const BorderRadius.vertical(
+              bottom: Radius.circular(4.0),
+            ),
+            side: BorderSide(
+              color: theme.scaffoldBackgroundColor,
+              width: 0.8,
+            ),
+          ),
+          color: theme.micaBackgroundColor,
+        ),
+        child: ValueListenableBuilder<TextEditingValue>(
+          valueListenable: controller,
+          builder: (context, value, _) {
+            final items =
+                AutoSuggestBox.defaultItemSorter(value.text, this.items);
+            late Widget result;
+            if (items.isEmpty) {
+              result = Padding(
+                padding: const EdgeInsets.only(bottom: 4.0),
+                child: _AutoSuggestBoxOverlayTile(
+                    text: localizations.noResultsFoundLabel),
+              );
+            } else {
+              result = ListView(
+                key: ValueKey<int>(items.length),
+                shrinkWrap: true,
+                padding: const EdgeInsets.only(bottom: 4.0),
+                children: List.generate(items.length, (index) {
+                  final item = items[index];
+                  return _AutoSuggestBoxOverlayTile(
+                    text: '$item',
+                    onSelected: () {
+                      onSelected(item);
+                    },
+                  );
+                }),
+              );
+            }
+            return result;
+          },
+        ),
+      ),
+    );
+  }
+}
+
+class _AutoSuggestBoxOverlayTile extends StatefulWidget {
+  const _AutoSuggestBoxOverlayTile({
+    Key? key,
+    required this.text,
+    this.onSelected,
+  }) : super(key: key);
+
+  final String text;
+  final VoidCallback? onSelected;
+
+  @override
+  __AutoSuggestBoxOverlayTileState createState() =>
+      __AutoSuggestBoxOverlayTileState();
+}
+
+class __AutoSuggestBoxOverlayTileState extends State<_AutoSuggestBoxOverlayTile>
+    with SingleTickerProviderStateMixin {
+  late AnimationController controller;
+
+  @override
+  void initState() {
+    super.initState();
+    controller = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 125),
+    );
+    controller.forward();
+  }
+
+  @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = FluentTheme.of(context);
+    return HoverButton(
+      onPressed: widget.onSelected,
+      margin: const EdgeInsets.only(top: 4.0, left: 4.0, right: 4.0),
+      builder: (context, states) => Stack(
+        children: [
+          Container(
+            height: 40.0,
+            padding: const EdgeInsets.symmetric(horizontal: 10.0),
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(6.0),
+              color: ButtonThemeData.uncheckedInputColor(
+                theme,
+                states.isDisabled ? {ButtonStates.none} : states,
+              ),
+            ),
+            alignment: Alignment.centerLeft,
+            child: EntrancePageTransition(
+              child: Text(
+                widget.text,
+                style: theme.typography.body,
+              ),
+              animation: Tween<double>(
+                begin: 0.75,
+                end: 1.0,
+              ).animate(CurvedAnimation(
+                parent: controller,
+                curve: Curves.easeOut,
+              )),
+              vertical: true,
+            ),
+          ),
+          if (states.isFocused)
+            Positioned(
+              top: 11.0,
+              bottom: 11.0,
+              left: 0.0,
+              child: Container(
+                width: 3.0,
+                decoration: BoxDecoration(
+                  color: theme.accentColor,
+                  borderRadius: BorderRadius.circular(4.0),
+                ),
+              ),
+            ),
+        ],
       ),
     );
   }